Create api_client.rb and move APIClient class from main.rb
